# docs/openapi-v1-draft.yaml
openapi: 3.0.3
info:
  title: Private Knowledge QA System API
  version: 0.1.0-current-backend
  description: >
    OpenAPI specification synchronized with the current Go backend implementation.
    This document intentionally reflects the actual runtime behavior, including
    placeholder SSE responses and endpoints that currently return 501 Not Implemented.
servers:
  - url: /api/v1
...
paths:
...
  /sessions:
    get:
      tags: [Sessions]
      summary: List sessions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/KeywordParam'
      responses:
        '200':
          description: Session list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    post:
      tags: [Sessions]
      summary: Create a new session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSessionResponse'
        '400':
          $ref: '#/components/responses/ValidationErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /sessions/{sessionId}:
    get:
      tags: [Sessions]
      summary: Get session detail with messages
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: Session detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
    delete:
      tags: [Sessions]
      summary: Soft delete a session
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: Session deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /sessions/{sessionId}/messages:
    post:
      tags: [Sessions]
      summary: Stream an assistant reply for a session
      description: >
        The backend now supports basic SSE chat for sessions that are not bound
        to a knowledge base. The current implementation uses DeepSeek as the
        provider, persists the user and assistant messages, and returns stream
        events in the order `meta`, `delta`, `done`. Provider failures are
        emitted as `event: error` inside the stream after the HTTP status has
        already been set to `200`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: SSE stream for the assistant reply
          content:
            text/event-stream:
              schema:
                type: string
                example: |-
                  event: meta
                  data: {"message_id":"9f45c628-8a1a-49ea-a714-e8354d4f3fdc","grounded":false,"model":"deepseek-chat"}

                  event: delta
                  data: {"content":"你好，"}

                  event: delta
                  data: {"content":"我可以帮你解释这个系统。"}

                  event: done
                  data: {"finish_reason":"stop"}
        '400':
          $ref: '#/components/responses/ValidationFailedResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '501':
          $ref: '#/components/responses/FeatureNotReadyResponse'
  /sessions/{sessionId}/messages/{messageId}/regenerate:
    post:
      tags: [Sessions]
      summary: Regenerate an existing assistant answer
      description: This endpoint currently returns 501 Not Implemented.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
        - $ref: '#/components/parameters/MessageIdParam'
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '501':
          $ref: '#/components/responses/FeatureNotReadyResponse'
  /sessions/{sessionId}/stream/stop:
    post:
      tags: [Sessions]
      summary: Stop current streaming task for the session
      description: This endpoint currently returns 501 Not Implemented.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '501':
          $ref: '#/components/responses/FeatureNotReadyResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    KeywordParam:
      name: keyword
      in: query
      schema:
        type: string
    SessionIdParam:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MessageIdParam:
      name: messageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    UnauthorizedResponse:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFoundResponse:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationErrorResponse:
      description: Validation failed or request body invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    FeatureNotReadyResponse:
      description: Feature not implemented in the current backend
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
      required: [field, message]
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 40000
        message:
          type: string
          example: validation failed
        data:
          nullable: true
        request_id:
          type: string
          example: 9b2daeed-adae-42ea-9ce2-7e60f22e9308
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
      required: [code, message, data]
    SimpleSuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          type: object
          additionalProperties: false
      required: [code, message, data]
...
    SessionSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        model:
          type: string
        knowledge_base_id:
          type: string
          format: uuid
          nullable: true
        knowledge_base_name:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, user_id, model, created_at, updated_at]
    SessionListData:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SessionSummary'
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
      required: [items, total, page, size]
    SessionListResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/SessionListData'
      required: [code, message, data]
    CreateSessionRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          nullable: true
        model:
          type: string
          description: Optional. If omitted, the backend uses `AI_DEFAULT_CHAT_MODEL`.
        knowledge_base_id:
          type: string
          format: uuid
          nullable: true
    SendMessageRequest:
      type: object
      required: [content]
      additionalProperties: false
      properties:
        content:
          type: string
    CreateSessionResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          required: [session_id]
          properties:
            session_id:
              type: string
              format: uuid
      required: [code, message, data]
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        reply_to_message_id:
          type: string
          format: uuid
          nullable: true
        content:
          type: string
        status:
          type: string
          enum: [completed, failed]
        model_used:
          type: string
          nullable: true
        grounded:
          type: boolean
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - session_id
        - role
        - content
        - status
        - grounded
        - prompt_tokens
        - completion_tokens
        - total_tokens
        - created_at
        - updated_at
    SessionDetailData:
      type: object
      properties:
        session:
          $ref: '#/components/schemas/SessionSummary'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
      required: [session, messages]
    SessionDetailResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/SessionDetailData'
      required: [code, message, data]
...